# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://notin.ir/api",
            "description": "Production Server"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "دریافت لیست دسته بندی‌ها",
                "description": "بازیابی لیست تمام دسته بندی‌ها با امکان جستجو",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "جستجو بر اساس نام دسته بندی",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "مدل‌های سه بعدی"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "شماره صفحه",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "لیست دسته بندی‌ها دریافت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "مدل‌های سه بعدی"
                                                    },
                                                    "color": {
                                                        "description": "رنگ دسته بندی",
                                                        "type": "string",
                                                        "example": "#FF5733"
                                                    },
                                                    "order": {
                                                        "description": "ترتیب نمایش",
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "دریافت جزئیات دسته بندی",
                "description": "بازیابی اطلاعات کامل یک دسته بندی و آمار محصولات آن",
                "operationId": "4cdf41756315109f1efc72787db5830a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه دسته بندی",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "اطلاعات دسته بندی دریافت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "مدل‌های سه بعدی"
                                        },
                                        "color": {
                                            "type": "string",
                                            "example": "#FF5733"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "توضیح دسته بندی"
                                        },
                                        "order": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00Z"
                                        },
                                        "total_likes": {
                                            "description": "کل لایک‌های تمام محصولات",
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "total_views": {
                                            "description": "کل مشاهدات تمام محصولات",
                                            "type": "integer",
                                            "example": 5000
                                        },
                                        "total_purchased": {
                                            "description": "کل خریدهای تمام محصولات",
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "total_3d_products": {
                                            "description": "تعداد محصولات سه بعدی",
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "total_paid_products": {
                                            "description": "تعداد محصولات پولی",
                                            "type": "integer",
                                            "example": 30
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "دسته بندی پیدا نشد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "دسته بندی یافت نشد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "",
                "bearerFormat": "",
                "scheme": "bearer"
            }
        }
    },
...
}